+++ src/main/java/histex/Histogram.java
@@ -9,10 +9,10 @@ public interface Histogram {
     }
   }
 
-  float getRank(float v);
+  double getRank(float v);
 
-  default float[] getRanks(float[] values) {
-    float[] result = new float[values.length];
+  default double[] getRanks(float[] values) {
+    double[] result = new double[values.length];
     for (int i=0; i<values.length; i++) {
       result[i] = getRank(values[i]);
     }
